Split latex functions into separate file.
